Initial commit
Initial commit contians python script, accompanying .json file and an Exe version for those who dont want to mess around with python.
The .json file is necessary if you use either the python script or the exe. It contains the DLC content lists and therefore it needs to be in the same folder as the python script or exe (whichever you are using).